ShaguPlates: add watchlist
A new watchlist feature has been added for superwow users. This allows players to overwrite
the nameplate color with a specific color when the unit is attacking a target which is registered
in the watchlist. As an example, this can be used for tanks to apply a specific color when a unit
is targeting off-tanks, so that they don't really need to worry about these units.

Three new settings has been added:
- A checkbox setting "Overwrite If Unit Is Attacking Someone On The Watchlist" to enable this feature.
- A new color setting "Overwrite If Unit Is Attackiung Someone On The Watchlist" to choose the desired color.
- A new list setting "Watchlist" to add player names to the watchlist.